# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task API",
    "description": "A simple and robust CRUD API for managing a To-Do list built with Express and Bun.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3020",
      "description": "Local Development Server"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "API Root Info",
        "description": "Returns basic metadata and endpoint summary for the API.",
        "responses": {
          "200": {
            "description": "API metadata returned successfully",
            "content": {
              "application/json": {
                "example": {
                  "name": "Task API",
                  "version": "1.0",
                  "endpoints": ["/tasks", "/stats", "/docs"]
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health Check",
        "description": "Returns server status to confirm service health.",
        "responses": {
          "200": {
            "description": "Server is healthy",
            "content": {
              "application/json": {
                "example": { "status": "ok" }
              }
            }
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "summary": "List tasks",
        "description": "Returns all tasks, with optional filtering by done status and keyword search.",
        "parameters": [
          {
            "name": "done",
            "in": "query",
            "description": "Filter by completed status (true or false)",
            "required": false,
            "schema": { "type": "boolean" }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Filter tasks whose title contains search term",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of tasks retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Task" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "description": "Creates a new task with title and sets done default to false.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string", "example": "Buy milk" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "400": {
            "description": "Invalid input - missing or empty title",
            "content": {
              "application/json": {
                "example": { "error": "Title is required and must be a non-empty string" }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get task by ID",
        "description": "Returns details of a single task by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Task ID",
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Task details retrieved",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "example": { "error": "Task 99 not found" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update task",
        "description": "Updates title and/or done status for a task.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Task ID",
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "example": "Buy whole milk" },
                  "done": { "type": "boolean", "example": true }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "400": {
            "description": "Invalid body payload",
            "content": {
              "application/json": {
                "example": { "error": "Provide title and/or done to update task" }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "example": { "error": "Task 99 not found" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete task",
        "description": "Removes a task by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Task ID",
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "204": {
            "description": "Task deleted successfully (No Content)"
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "example": { "error": "Task 99 not found" }
              }
            }
          }
        }
      }
    },
    "/stats": {
      "get": {
        "summary": "Task statistics",
        "description": "Returns counts for total, done, and open tasks.",
        "responses": {
          "200": {
            "description": "Task stats calculated",
            "content": {
              "application/json": {
                "example": { "total": 3, "done": 1, "open": 2 }
              }
            }
          }
        }
      }
    },
    "/reset": {
      "post": {
        "summary": "Reset tasks list",
        "description": "Resets memory store back to initial 3 sample tasks.",
        "responses": {
          "200": {
            "description": "Tasks reset completed",
            "content": {
              "application/json": {
                "example": {
                  "message": "Tasks reset to initial state",
                  "tasks": [
                    { "id": 1, "title": "Learn Express & Bun", "done": true },
                    { "id": 2, "title": "Build CRUD API", "done": false },
                    { "id": 3, "title": "Setup Swagger UI", "done": false }
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "title": { "type": "string", "example": "Learn Express & Bun" },
          "done": { "type": "boolean", "example": true }
        }
      }
    }
  }
}